Get civic service request data

Get civic service request data from a registered open311 endpoint. o311_request queries a single service request by ID. o311_requests queries a single page of service requests. o311_request_all tries to iterate through all pages of an endpoint to return a complete dataset of service requests.

Usage

o311_requests(
  service_code = NULL,
  start_date = NULL,
  end_date = NULL,
  status = NULL,
  page = NULL,
  ...
)

o311_request(service_request_id, ...)

o311_request_all(
  service_code = NULL,
  start_date = NULL,
  end_date = NULL,
  status = NULL,
  ...,
  max_pages = Inf,
  progress = TRUE
)

Arguments

service_code

[character]

IDs of the service types to be queried. Defaults to all available codes of an endpoint. A list of all available service codes can be retrieved using o311_services.

start_date, end_date

[POSIXt]

Start date and end date of the query results. Must be date-time objects. If not specified, defaults to the last 90 days.

status

[character]

Status of the public service ticket. Can be one of "open" or "closed". If NULL, returns all types of tickets.

page

[integer]

Page of the response. Most endpoints paginate their responses in a way that only a limited number of tickets are returned with each query. To retrieve all data, consider using o311_request_all.

...

Further endpoint-specific parameters as documented in the respective endpoint reference.

service_request_id

[character]

Identifier of a single service request. Request IDs can usually be retrieved from o311_requests.

max_pages

[integer]

Number of pages to search until the result is returned.

progress

[logical]

Whether to show a waiter indicating the current page iteration.

Value

A dataframe containing data on civic service requests. The dataframe can contain varying columns depending on the open311 implementation.

Details

o311_request_all applies a number of checks to determine when to stop searching. First, many endpoints return an error if the last page is exceeded. Thus, if the last page request failed, break. Second, if exceeding the pagination limit does not return an error, the response is compared with the previous response. If identical, the response is discarded and all previous responses returned. Finally, if the page exceeds max_pages, the responses up to this point are returned.

open311 leaves space for endpoints to implement their own request parameters. These parameters can be provided using dot arguments. These arguments are not validated or pre-processed. Date-time objects must be formatted according to the w3c standard. Some more common parameters include:

q: Perform a text search across all requests.

update_after/updated_before: Limit request according to request update dates.

per_page: Specifiy the maximum number of requests per page.

extensions: Adds a nested attribute "extended_attributes" to the response.

long/lat/radius: Searches for requests in a fixed radius around a coordinate.

As dot arguments deviate from the open311 standard, they are not guaranteed to be available for every endpoint and might be removed without further notice. Refer to the endpoint docs to learn more about custom parameters (o311_endpoints()$docs).

See also

Examples

o311_api("zurich")

can_connect <- o311_ok()
if (can_connect) {
  # retrieve requests from the last two days
  now <- Sys.time()
  two_days <- 60 * 60 * 24 * 2
  o311_requests(end_date = now, start_date = now - two_days)

  # retrieve only open tickets
  tickets <- o311_requests(status = "open")

  # request the first ticket of the previous response
  rid <- as.character(tickets$service_request_id[1])
  o311_request(rid)

  # request all data
  o311_request_all()

  # request data of the first 5 pages
  o311_request_all(max_pages = 5)
}
#> Receiving page 4...Receiving page 5...Receiving page 6...Receiving page 7...Receiving page 8...Receiving page 9...Receiving page 10...Receiving page 11...Receiving page 12...Receiving page 13...Receiving page 14...Receiving page 15...Receiving page 16...Receiving page 17...Receiving page 18...Receiving page 19...Receiving page 20...Receiving page 21...Receiving page 22...Receiving page 23...Receiving page 24...Receiving page 25...Receiving page 26...Receiving page 27...Receiving page 28...Receiving page 29...Receiving page 30...Receiving page 31...Receiving page 32...Receiving page 33...Receiving page 34...Receiving page 35...Receiving page 36...Receiving page 37...Receiving page 38...Receiving page 39...Receiving page 40...Receiving page 41...Receiving page 42...Receiving page 43...Receiving page 44...Receiving page 45...Receiving page 46...Receiving page 47...Receiving page 48...Receiving page 49...Receiving page 50...Receiving page 51...Receiving page 52...Receiving page 53...Receiving page 54...Receiving page 55...Receiving page 56...Receiving page 57...Receiving page 58...Receiving page 59...Receiving page 60...Receiving page 61...Receiving page 62...Receiving page 63...Receiving page 64...Receiving page 65...Receiving page 66...Receiving page 67...Receiving page 68...Receiving page 69...Receiving page 70...Receiving page 71...Receiving page 72...Receiving page 73...Receiving page 74...Receiving page 75...Receiving page 76...Receiving page 77...Receiving page 78...Receiving page 79...Receiving page 80...Receiving page 81...Receiving page 82...Receiving page 83...Receiving page 84...Receiving page 85...Receiving page 86...Receiving page 87...Receiving page 88...Receiving page 89...Receiving page 90...Receiving page 91...Receiving page 92...Receiving page 93...Receiving page 94...Receiving page 95...Receiving page 96...Receiving page 97...Receiving page 98...Receiving page 99...Receiving page 100...Receiving page 101...Receiving page 102...Receiving page 103...Receiving page 104...Receiving page 105...Receiving page 106...Receiving page 107...Receiving page 108...Receiving page 109...Receiving page 110...Receiving page 111...Receiving page 112...Receiving page 113...Receiving page 114...Receiving page 115...Receiving page 116...Receiving page 117...Receiving page 118...Receiving page 119...Receiving page 120...Receiving page 121...Receiving page 122...Receiving page 123...Receiving page 124...Receiving page 125...Receiving page 126...Receiving page 127...Receiving page 128...Receiving page 129...Receiving page 130...Receiving page 131...Receiving page 132...Receiving page 133...Receiving page 134...Receiving page 135...Receiving page 136...Receiving page 137...Receiving page 138...Receiving page 139...Receiving page 140...Receiving page 141...Receiving page 142...Receiving page 143...Receiving page 144...Receiving page 145...Receiving page 146...Receiving page 147...Receiving page 148...Receiving page 149...Receiving page 150...Receiving page 151...Receiving page 152...Receiving page 153...Receiving page 154...Receiving page 155...Receiving page 156...Receiving page 157...Receiving page 158...Receiving page 159...Receiving page 160...Receiving page 161...Receiving page 162...Receiving page 163...Receiving page 164...Receiving page 165...Receiving page 166...Receiving page 167...Receiving page 168...Receiving page 169...Receiving page 170...Receiving page 171...Receiving page 172...Receiving page 173...Receiving page 174...Receiving page 175...Receiving page 176...Receiving page 177...Receiving page 178...Receiving page 179...Receiving page 180...Receiving page 181...Receiving page 182...Receiving page 183...Receiving page 184...Receiving page 185...Receiving page 186...Receiving page 187...Receiving page 188...Receiving page 189...Receiving page 190...Receiving page 191...Receiving page 192...Receiving page 193...Receiving page 194...Receiving page 195...Receiving page 196...Receiving page 197...Receiving page 198...Receiving page 199...Receiving page 200...Receiving page 201...Receiving page 202...Receiving page 203...Receiving page 204...Receiving page 205...Receiving page 206...Receiving page 207...Receiving page 208...Receiving page 209...Receiving page 210...Receiving page 211...Receiving page 212...Receiving page 213...Receiving page 214...Receiving page 215...Receiving page 216...Receiving page 217...Receiving page 218...Receiving page 219...Receiving page 220...Receiving page 221...Receiving page 222...Receiving page 223...Receiving page 224...Receiving page 225...Receiving page 226...Receiving page 227...Receiving page 228...Receiving page 229...Receiving page 230...Receiving page 231...Receiving page 232...Receiving page 233...Receiving page 234...Receiving page 235...Receiving page 236...Receiving page 237...Receiving page 238...Receiving page 239...Receiving page 240...Receiving page 241...Receiving page 242...Receiving page 243...Receiving page 244...Receiving page 245...Receiving page 246...Receiving page 247...Receiving page 248...Receiving page 249...Receiving page 250...Receiving page 251...Receiving page 252...Receiving page 253...Receiving page 254...Receiving page 255...Receiving page 256...Receiving page 257...Receiving page 258...Receiving page 259...Receiving page 260...Receiving page 261...Receiving page 262...Receiving page 263...Receiving page 264...Receiving page 265...Receiving page 266...Receiving page 267...Receiving page 268...Receiving page 269...Receiving page 270...Receiving page 271...Receiving page 272...Receiving page 273...Receiving page 274...Receiving page 275...Receiving page 276...Receiving page 277...Receiving page 278...Receiving page 279...Receiving page 280...Receiving page 281...Receiving page 282...Receiving page 283...Receiving page 284...Receiving page 285...Receiving page 286...Receiving page 287...Receiving page 288...Receiving page 289...Receiving page 290...Receiving page 291...Receiving page 292...Receiving page 293...Receiving page 294...Receiving page 295...Receiving page 296...Receiving page 297...Receiving page 298...Receiving page 299...Receiving page 300...Receiving page 301...Receiving page 302...Receiving page 303...Receiving page 304...Receiving page 305...Receiving page 306...Receiving page 307...Receiving page 308...Receiving page 309...Receiving page 310...Receiving page 311...Receiving page 312...Receiving page 313...Receiving page 314...Receiving page 315...Receiving page 316...Receiving page 317...Receiving page 318...Receiving page 319...Receiving page 320...Receiving page 321...Receiving page 322...Receiving page 323...Receiving page 324...Receiving page 325...Receiving page 326...Receiving page 327...Receiving page 328...Receiving page 329...Receiving page 330...Receiving page 331...Receiving page 332...Receiving page 333...Receiving page 334...Receiving page 335...Receiving page 336...Receiving page 337...Receiving page 338...Receiving page 339...Receiving page 340...Receiving page 341...Receiving page 342...Receiving page 343...Receiving page 344...Receiving page 345...Receiving page 346...Receiving page 347...Receiving page 348...Receiving page 349...Receiving page 350...Receiving page 351...Receiving page 352...Receiving page 353...Receiving page 354...Receiving page 355...Receiving page 356...Receiving page 357...Receiving page 358...Receiving page 359...Receiving page 360...Receiving page 361...Receiving page 362...Receiving page 363...Receiving page 364...Receiving page 365...Receiving page 366...Receiving page 367...Receiving page 368...Receiving page 369...Receiving page 370...Receiving page 371...Receiving page 372...Receiving page 373...Receiving page 374...Receiving page 375...Receiving page 376...Receiving page 377...Receiving page 378...Receiving page 379...Receiving page 380...Receiving page 381...Receiving page 382...Receiving page 383...Receiving page 384...Receiving page 385...Receiving page 386...Receiving page 387...Receiving page 388...Receiving page 389...Receiving page 390...Receiving page 391...Receiving page 392...Receiving page 393...Receiving page 394...Receiving page 395...Receiving page 396...Receiving page 397...Receiving page 398...Receiving page 399...Receiving page 400...Receiving page 401...Receiving page 402...Receiving page 403...Receiving page 404...Receiving page 405...Receiving page 406...Receiving page 407...Receiving page 408...Receiving page 409...Receiving page 410...Receiving page 411...Receiving page 412...Receiving page 413...Receiving page 414...Receiving page 415...Receiving page 416...Receiving page 417...Receiving page 418...Receiving page 419...Receiving page 420...Receiving page 421...Receiving page 422...Receiving page 423...Receiving page 424...Receiving page 425...Receiving page 426...Receiving page 427...Receiving page 428...Receiving page 429...Receiving page 430...Receiving page 431...Receiving page 432...Receiving page 433...Receiving page 434...Receiving page 435...Receiving page 436...Receiving page 437...Receiving page 438...Receiving page 439...Receiving page 440...Receiving page 441...Receiving page 442...Receiving page 443...Receiving page 444...Receiving page 445...Receiving page 446...Receiving page 447...Receiving page 448...Receiving page 449...Receiving page 450...Receiving page 451...Receiving page 452...Receiving page 453...Receiving page 454...Receiving page 455...Receiving page 456...Receiving page 457...Receiving page 458...Receiving page 459...Receiving page 460...Receiving page 461...Receiving page 462...Receiving page 463...Receiving page 464...Receiving page 465...Receiving page 466...Receiving page 467...Receiving page 468...Receiving page 469...Receiving page 470...Receiving page 471...Receiving page 472...Receiving page 473...Receiving page 474...Receiving page 475...Receiving page 476...Receiving page 477...Receiving page 478...Receiving page 479...Receiving page 480...Receiving page 481...Receiving page 482...Receiving page 483...Receiving page 484...Receiving page 485...Receiving page 486...Receiving page 487...Receiving page 488...Receiving page 489...Receiving page 490...Receiving page 491...Receiving page 492...Receiving page 493...Receiving page 494...Receiving page 495...Receiving page 496...Receiving page 497...Receiving page 498...Receiving page 499...Receiving page 500...Receiving page 501...Receiving page 502...Receiving page 503...Receiving page 504...Receiving page 505...Receiving page 506...Receiving page 507...Receiving page 508...Receiving page 509...Receiving page 510...Receiving page 511...Receiving page 512...Receiving page 513...Receiving page 514...Receiving page 515...Receiving page 516...Receiving page 517...Receiving page 518...Receiving page 519...Receiving page 520...Receiving page 521...Receiving page 522...Receiving page 523...Receiving page 524...Receiving page 525...Receiving page 526...Receiving page 527...Receiving page 528...Receiving page 529...Receiving page 530...Receiving page 531...Receiving page 532...Receiving page 533...Receiving page 534...Receiving page 535...Receiving page 536...Receiving page 537...Receiving page 538...Receiving page 539...Receiving page 540...Receiving page 541...Receiving page 542...Receiving page 543...Receiving page 544...Receiving page 545...Receiving page 546...Receiving page 547...Receiving page 548...Receiving page 549...Receiving page 550...Receiving page 551...Receiving page 552...Receiving page 553...Receiving page 554...Receiving page 555...Receiving page 556...Receiving page 557...Receiving page 558...Receiving page 559...Receiving page 560...Receiving page 561...Receiving page 562...Receiving page 563...Receiving page 564...Receiving page 565...Receiving page 566...Receiving page 567...Receiving page 568...Receiving page 569...Receiving page 570...Receiving page 571...Receiving page 572...Receiving page 573...Receiving page 574...Receiving page 575...Receiving page 576...Receiving page 577...Receiving page 578...Receiving page 579...Receiving page 580...Receiving page 581...Receiving page 582...Receiving page 583...Receiving page 584...Receiving page 585...Receiving page 586...Receiving page 587...Receiving page 588...Receiving page 589...Receiving page 590...Receiving page 591...Receiving page 592...Receiving page 593...Receiving page 594...Receiving page 595...Receiving page 596...Receiving page 597...Receiving page 598...Receiving page 599...Receiving page 600...Receiving page 601...Receiving page 602...Receiving page 603...Receiving page 604...Receiving page 605...Receiving page 606...Receiving page 607...Receiving page 608...Receiving page 609...Receiving page 610...Receiving page 611...Receiving page 612...Receiving page 613...Receiving page 614...Receiving page 615...Receiving page 616...Receiving page 617...Receiving page 618...Receiving page 619...Receiving page 620...Receiving page 621...Receiving page 622...Receiving page 623...Receiving page 624...Receiving page 625...Receiving page 626...Receiving page 627...Receiving page 628...Receiving page 629...Receiving page 630...Receiving page 631...Receiving page 632...Receiving page 633...Receiving page 634...Receiving page 635...Receiving page 636...Receiving page 637...Receiving page 638...Receiving page 639...Receiving page 640...Receiving page 641...Receiving page 642...Receiving page 643...Receiving page 644...Receiving page 645...Receiving page 646...Receiving page 647...Receiving page 648...Receiving page 649...Receiving page 650...Receiving page 651...Receiving page 652...Receiving page 653...Receiving page 654...Receiving page 655...Receiving page 656...Receiving page 657...Receiving page 658...Receiving page 659...Receiving page 660...Receiving page 661...Receiving page 662...Receiving page 663...Receiving page 664...Receiving page 665...Receiving page 666...Receiving page 667...Receiving page 668...Receiving page 669...Receiving page 670...Receiving page 671...Receiving page 672...Receiving page 673...Receiving page 674...Receiving page 675...Receiving page 676...Receiving page 677...Receiving page 678...Receiving page 679...Receiving page 680...Receiving page 681...Receiving page 682...Receiving page 683...Receiving page 684...Receiving page 685...Receiving page 686...Receiving page 687...Receiving page 688...Receiving page 689...Receiving page 690...Receiving page 691...Receiving page 692...Receiving page 693...Receiving page 694...Receiving page 695...Receiving page 696...Receiving page 697...Receiving page 698...Receiving page 699...Receiving page 700...Receiving page 701...Receiving page 702...Receiving page 703...Receiving page 704...Receiving page 705...Receiving page 706...Receiving page 707...Receiving page 708...Receiving page 709...Receiving page 710...Receiving page 711...Receiving page 712...Receiving page 713...Receiving page 714...Receiving page 715...Receiving page 716...Receiving page 717...Receiving page 718...Receiving page 719...Receiving page 720...Receiving page 721...Receiving page 722...Receiving page 723...Receiving page 724...Receiving page 725...Receiving page 726...Receiving page 727...Receiving page 728...Receiving page 729...Receiving page 730...Receiving page 731...Receiving page 732...Receiving page 733...Receiving page 734...Receiving page 735...Receiving page 736...Receiving page 737...Receiving page 738...Receiving page 739...Receiving page 740...Receiving page 741...Receiving page 742...Receiving page 743...Receiving page 744...Receiving page 745...Receiving page 746...Receiving page 747...Receiving page 748...Receiving page 749...Receiving page 750...Receiving page 751...Receiving page 752...Receiving page 753...Receiving page 754...Receiving page 755...Receiving page 756...Receiving page 757...Receiving page 758...Receiving page 759...Receiving page 760...Receiving page 761...Receiving page 762...Receiving page 763...Receiving page 764...Receiving page 765...Receiving page 766...Receiving page 767...Receiving page 768...Receiving page 769...Receiving page 770...Receiving page 771...Receiving page 772...Receiving page 773...Receiving page 774...Receiving page 775...Receiving page 776...Receiving page 777...Receiving page 778...Receiving page 779...Receiving page 780...Receiving page 781...Receiving page 782...Receiving page 783...Receiving page 784...Receiving page 785...Receiving page 786...Receiving page 787...Receiving page 788...Receiving page 789...Receiving page 790...Receiving page 791...Receiving page 792...Receiving page 793...Receiving page 794...Receiving page 795...Receiving page 796...Receiving page 797...Receiving page 4 out of 5...Receiving page 5 out of 5...
#> Simple feature collection with 5000 features and 13 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 8.465915 ymin: 47.32471 xmax: 8.599787 ymax: 47.43141
#> Geodetic CRS:  WGS 84
#> # A tibble: 5,000 × 14
#>    description       service_code updated_datetime service_notice interface_used
#>    <chr>             <chr>        <chr>            <chr>          <chr>         
#>  1 Hakenkreuzgrafit… Graffiti     2024-07-10T17:4…  NA            mobile        
#>  2 Hopp fcz: Hopp f… Graffiti     2024-07-10T17:4…  NA            iOS           
#>  3 Donnerstag um 13… Abfall/Samm… 2024-07-10T14:3… "Entsorgung +… iOS           
#>  4 Pfeil bei Tafel … Signalisati… 2024-07-10T14:4… "Guten Tag\r\… iOS           
#>  5 Tisch wurde leid… Abfall/Samm… 2024-07-10T14:3… "Entsorgung +… iOS           
#>  6 Holzabfall: Holz… Abfall/Samm… 2024-07-10T13:3… "Entsorgung +… iOS           
#>  7 Abfall: Abfall    Abfall/Samm… 2024-07-10T13:3… "Entsorgung +… iOS           
#>  8 Kübel: Kübel      Abfall/Samm… 2024-07-10T13:2… "Entsorgung +… iOS           
#>  9 Bretter: Bretter  Abfall/Samm… 2024-07-10T13:2… "Entsorgung +… iOS           
#> 10 Tisch: Tisch      Abfall/Samm… 2024-07-10T13:2… "Entsorgung +… desktop       
#> # ℹ 4,990 more rows
#> # ℹ 9 more variables: status <chr>, requested_datetime <chr>, title <chr>,
#> #   detail <chr>, service_request_id <int>, service_name <chr>,
#> #   agency_sent_datetime <chr>, media_url <chr>, geometry <POINT [°]>