Get civic service request data

Get civic service request data from a registered open311 endpoint. o311_request queries a single service request by ID. o311_requests queries a single page of service requests. o311_request_all tries to iterate through all pages of an endpoint to return a complete dataset of service requests.

Usage

o311_requests(
  service_code = NULL,
  start_date = NULL,
  end_date = NULL,
  status = NULL,
  page = NULL,
  ...
)

o311_request(service_request_id, ...)

o311_request_all(
  service_code = NULL,
  start_date = NULL,
  end_date = NULL,
  status = NULL,
  ...,
  max_pages = Inf,
  progress = TRUE
)

Arguments

service_code

[character]

IDs of the service types to be queried. Defaults to all available codes of an endpoint. A list of all available service codes can be retrieved using o311_services.

start_date, end_date

[POSIXt]

Start date and end date of the query results. Must be date-time objects. If not specified, defaults to the last 90 days.

status

[character]

Status of the public service ticket. Can be one of "open" or "closed". If NULL, returns all types of tickets.

page

[integer]

Page of the response. Most endpoints paginate their responses in a way that only a limited number of tickets are returned with each query. To retrieve all data, consider using o311_request_all.

...

Further endpoint-specific parameters as documented in the respective endpoint reference.

service_request_id

[character]

Identifier of a single service request. Request IDs can usually be retrieved from o311_requests.

max_pages

[integer]

Number of pages to search until the result is returned.

progress

[logical]

Whether to show a waiter indicating the current page iteration.

Value

A dataframe containing data on civic service requests. The dataframe can contain varying columns depending on the open311 implementation.

Details

o311_request_all applies a number of checks to determine when to stop searching. First, many endpoints return an error if the last page is exceeded. Thus, if the last page request failed, break. Second, if exceeding the pagination limit does not return an error, the response is compared with the previous response. If identical, the response is discarded and all previous responses returned. Finally, if the page exceeds max_pages, the responses up to this point are returned.

open311 leaves space for endpoints to implement their own request parameters. These parameters can be provided using dot arguments. These arguments are not validated or pre-processed. Date-time objects must be formatted according to the w3c standard. Some more common parameters include:

q: Perform a text search across all requests.

update_after/updated_before: Limit request according to request update dates.

per_page: Specifiy the maximum number of requests per page.

extensions: Adds a nested attribute "extended_attributes" to the response.

long/lat/radius: Searches for requests in a fixed radius around a coordinate.

As dot arguments deviate from the open311 standard, they are not guaranteed to be available for every endpoint and might be removed without further notice. Refer to the endpoint docs to learn more about custom parameters (o311_endpoints()$docs).

See also

Examples

o311_api("zurich")

can_connect <- o311_ok()
if (can_connect) {
  # retrieve requests from the last two days
  now <- Sys.time()
  two_days <- 60 * 60 * 24 * 2
  o311_requests(end_date = now, start_date = now - two_days)

  # retrieve only open tickets
  tickets <- o311_requests(status = "open")

  # request the first ticket of the previous response
  rid <- as.character(tickets$service_request_id[1])
  o311_request(rid)

  # request all data
  o311_request_all()

  # request data of the first 5 pages
  o311_request_all(max_pages = 5)
}
#> Receiving page 4...Receiving page 5...Receiving page 6...Receiving page 7...Receiving page 8...Receiving page 9...Receiving page 10...Receiving page 11...Receiving page 12...Receiving page 13...Receiving page 14...Receiving page 15...Receiving page 16...Receiving page 17...Receiving page 18...Receiving page 19...Receiving page 20...Receiving page 21...Receiving page 22...Receiving page 23...Receiving page 24...Receiving page 25...Receiving page 26...Receiving page 27...Receiving page 28...Receiving page 29...Receiving page 30...Receiving page 31...Receiving page 32...Receiving page 33...Receiving page 34...Receiving page 35...Receiving page 36...Receiving page 37...Receiving page 38...Receiving page 39...Receiving page 40...Receiving page 41...Receiving page 42...Receiving page 43...Receiving page 44...Receiving page 45...Receiving page 46...Receiving page 47...Receiving page 48...Receiving page 49...Receiving page 50...Receiving page 51...Receiving page 52...Receiving page 53...Receiving page 54...Receiving page 55...Receiving page 56...Receiving page 57...Receiving page 58...Receiving page 59...Receiving page 60...Receiving page 61...Receiving page 62...Receiving page 63...Receiving page 64...Receiving page 65...Receiving page 66...Receiving page 67...Receiving page 68...Receiving page 69...Receiving page 70...Receiving page 71...Receiving page 72...Receiving page 73...Receiving page 74...Receiving page 75...Receiving page 76...Receiving page 77...Receiving page 78...Receiving page 79...Receiving page 80...Receiving page 81...Receiving page 82...Receiving page 83...Receiving page 84...Receiving page 85...Receiving page 86...Receiving page 87...Receiving page 88...Receiving page 89...Receiving page 90...Receiving page 91...Receiving page 92...Receiving page 93...Receiving page 94...Receiving page 95...Receiving page 96...Receiving page 97...Receiving page 98...Receiving page 99...Receiving page 100...Receiving page 101...Receiving page 102...Receiving page 103...Receiving page 104...Receiving page 105...Receiving page 106...Receiving page 107...Receiving page 108...Receiving page 109...Receiving page 110...Receiving page 111...Receiving page 112...Receiving page 113...Receiving page 114...Receiving page 115...Receiving page 116...Receiving page 117...Receiving page 118...Receiving page 119...Receiving page 120...Receiving page 121...Receiving page 122...Receiving page 123...Receiving page 124...Receiving page 125...Receiving page 126...Receiving page 127...Receiving page 128...Receiving page 129...Receiving page 130...Receiving page 131...Receiving page 132...Receiving page 133...Receiving page 134...Receiving page 135...Receiving page 136...Receiving page 137...Receiving page 138...Receiving page 139...Receiving page 140...Receiving page 141...Receiving page 142...Receiving page 143...Receiving page 144...Receiving page 145...Receiving page 146...Receiving page 147...Receiving page 148...Receiving page 149...Receiving page 150...Receiving page 151...Receiving page 152...Receiving page 153...Receiving page 154...Receiving page 155...Receiving page 156...Receiving page 157...Receiving page 158...Receiving page 159...Receiving page 160...Receiving page 161...Receiving page 162...Receiving page 163...Receiving page 164...Receiving page 165...Receiving page 166...Receiving page 167...Receiving page 168...Receiving page 169...Receiving page 170...Receiving page 171...Receiving page 172...Receiving page 173...Receiving page 174...Receiving page 175...Receiving page 176...Receiving page 177...Receiving page 178...Receiving page 179...Receiving page 180...Receiving page 181...Receiving page 182...Receiving page 183...Receiving page 184...Receiving page 185...Receiving page 186...Receiving page 187...Receiving page 188...Receiving page 189...Receiving page 190...Receiving page 191...Receiving page 192...Receiving page 193...Receiving page 194...Receiving page 195...Receiving page 196...Receiving page 197...Receiving page 198...Receiving page 199...Receiving page 200...Receiving page 201...Receiving page 202...Receiving page 203...Receiving page 204...Receiving page 205...Receiving page 206...Receiving page 207...Receiving page 208...Receiving page 209...Receiving page 210...Receiving page 211...Receiving page 212...Receiving page 213...Receiving page 214...Receiving page 215...Receiving page 216...Receiving page 217...Receiving page 218...Receiving page 219...Receiving page 220...Receiving page 221...Receiving page 222...Receiving page 223...Receiving page 224...Receiving page 225...Receiving page 226...Receiving page 227...Receiving page 228...Receiving page 229...Receiving page 230...Receiving page 231...Receiving page 232...Receiving page 233...Receiving page 234...Receiving page 235...Receiving page 236...Receiving page 237...Receiving page 238...Receiving page 239...Receiving page 240...Receiving page 241...Receiving page 242...Receiving page 243...Receiving page 244...Receiving page 245...Receiving page 246...Receiving page 247...Receiving page 248...Receiving page 249...Receiving page 250...Receiving page 251...Receiving page 252...Receiving page 253...Receiving page 254...Receiving page 255...Receiving page 256...Receiving page 257...Receiving page 258...Receiving page 259...Receiving page 260...Receiving page 261...Receiving page 262...Receiving page 263...Receiving page 264...Receiving page 265...Receiving page 266...Receiving page 267...Receiving page 268...Receiving page 269...Receiving page 270...Receiving page 271...Receiving page 272...Receiving page 273...Receiving page 274...Receiving page 275...Receiving page 276...Receiving page 277...Receiving page 278...Receiving page 279...Receiving page 280...Receiving page 281...Receiving page 282...Receiving page 283...Receiving page 284...Receiving page 285...Receiving page 4 out of 5...Receiving page 5 out of 5...
#> Simple feature collection with 5000 features and 13 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 8.465915 ymin: 47.32471 xmax: 8.599787 ymax: 47.43141
#> Geodetic CRS:  WGS 84
#> # A tibble: 5,000 × 14
#>    title agency_sent_datetime service_request_id service_code description status
#>    <chr> <chr>                             <int> <chr>        <chr>       <chr> 
#>  1 Donn… 2024-07-10T13:57:04…              58425 Abfall/Samm… Donnerstag… closed
#>  2 Pfei… 2024-07-10T13:42:04…              58424 Signalisati… Pfeil bei … closed
#>  3 Tisc… 2024-07-10T12:52:04…              58423 Abfall/Samm… Tisch wurd… closed
#>  4 Holz… 2024-07-10T12:52:04…              58422 Abfall/Samm… Holzabfall… closed
#>  5 Abfa… 2024-07-10T12:47:04…              58421 Abfall/Samm… Abfall: Ab… closed
#>  6 Kübel 2024-07-10T12:47:04…              58420 Abfall/Samm… Kübel: Küb… closed
#>  7 Bret… 2024-07-10T12:47:04…              58419 Abfall/Samm… Bretter: B… closed
#>  8 Tisch 2024-07-10T12:47:04…              58418 Abfall/Samm… Tisch: Tis… closed
#>  9 Sign… 2024-07-10T11:07:04…              58416 Signalisati… Signaltafe… closed
#> 10 Zwei… 2024-07-10T15:27:05…              58415 Allgemein    Zweirad: S… open  
#> # ℹ 4,990 more rows
#> # ℹ 8 more variables: interface_used <chr>, service_name <chr>,
#> #   media_url <chr>, service_notice <chr>, requested_datetime <chr>,
#> #   updated_datetime <chr>, detail <chr>, geometry <POINT [°]>